Struct rusoto_pricing::PricingClient

pub struct PricingClient { /* fields omitted */ }

A client for the AWS Pricing API.

Methods

impl PricingClient

pub fn new(region: Region) -> PricingClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> PricingClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,