Trait rusoto_pricing::Pricing

pub trait Pricing {
    fn describe_services(
        &self, 
        input: DescribeServicesRequest
    ) -> RusotoFuture<DescribeServicesResponse, DescribeServicesError>;

    fn get_attribute_values(
        &self, 
        input: GetAttributeValuesRequest
    ) -> RusotoFuture<GetAttributeValuesResponse, GetAttributeValuesError>;

    fn get_products(
        &self, 
        input: GetProductsRequest
    ) -> RusotoFuture<GetProductsResponse, GetProductsError>;
}

Trait representing the capabilities of the AWS Pricing API. AWS Pricing clients implement this trait.

Required methods

fn describe_services(
&self,
input: DescribeServicesRequest
) -> RusotoFuture<DescribeServicesResponse, DescribeServicesError>

Returns the metadata for one service or a list of the metadata for all services. Use this without a service code to get the service codes for all services. Use it with a service code, such as AmazonEC2, to get information specific to that service, such as the attribute names available for that service. For example, some of the attribute names available for EC2 are volumeType, maxIopsVolume, operation, locationType, and instanceCapacity10xlarge.

fn get_attribute_values(
&self,
input: GetAttributeValuesRequest
) -> RusotoFuture<GetAttributeValuesResponse, GetAttributeValuesError>

Returns a list of attribute values. Attibutes are similar to the details in a Price List API offer file. For a list of available attributes, see Offer File Definitions in the AWS Billing and Cost Management User Guide.

fn get_products(
&self,
input: GetProductsRequest
) -> RusotoFuture<GetProductsResponse, GetProductsError>

Returns a list of all products that match the filter criteria.

Loading content...